Refactored torch_training_utils checkpointing functions
