+++ applications/harnesses/ch_DDP_loderunner/train_LodeRunner_ddp.py
@@ -10,6 +10,8 @@ from torch.nn.parallel import DistributedDataParallel as DDP
 from yoke.models.vit.swin.bomberman import LodeRunner
 from yoke.datasets.lsc_dataset import LSC_rho2rho_temporal_DataSet
 import yoke.torch_training_utils as tr
+from yoke.utils.checkpointing import load_model_and_optimizer
+from yoke.utils.checkpointing import save_model_and_optimizer
 from yoke.lr_schedulers import CosineWithWarmupScheduler
 from yoke.helpers import cli
 
@@ -176,7 +178,7 @@ def main(args, rank, world_size, local_rank, device):
     # Wait to move model to GPU until after the checkpoint load. Then
     # explicitly move model and optimizer state to GPU.
     if CONTINUATION:
-        model, starting_epoch = tr.load_model_and_optimizer(
+        model, starting_epoch = load_model_and_optimizer(
             checkpoint,
             optimizer,
             available_models,
@@ -314,11 +316,11 @@ def main(args, rank, world_size, local_rank, device):
     chkpt_name_str = "study{0:03d}_modelState_epoch{1:04d}.pth"
     new_chkpt_path = os.path.join("./", chkpt_name_str.format(studyIDX, epochIDX))
 
-    tr.save_model_and_optimizer(
-        model, 
-        optimizer, 
+    save_model_and_optimizer(
+        model,
+        optimizer,
         epochIDX,
-        new_chkpt_path, 
+        new_chkpt_path,
         model_class=LodeRunner,
         model_args=model_args
     )
